Split Value and ValueRef into separate modules
